Добавление вопроса

+++ Admin_Module/Form1.cs
@@ -38,7 +38,7 @@ namespace Admin_Module
 				{
 					UseHeaderRow = false
 				}
-			}) ;
+			});
 			tableCollection = db.Tables;
 			DataTable table = tableCollection[Convert.ToString(tableCollection[0].TableName)];
 			dataGridView1.DataSource = table;
@@ -63,8 +63,8 @@ namespace Admin_Module
 		private void toolStripButton1_Click(object sender, EventArgs e)
 		{
 			dataGridView1.Visible = false;
-			toolStripButton3.Enabled = false;
-			toolStripButton2.Enabled = false;
+			btn_editor.Enabled = false;
+			btn_newquestion.Enabled = false;
 			button1.Enabled = false;
 			button2.Enabled = false;
 			textBox1.Enabled = false;
@@ -81,8 +81,9 @@ namespace Admin_Module
 					Text = filename;
 					OpenExcelFile(filename);
 				}
-				else throw new Exception("Файл не выбран");
-				toolStripButton2.Enabled = true;
+				else throw new Exception("Файл не был сохранен");
+				btn_editor.Enabled = true;
+				btn_newquestion.Enabled = true;
 				button1.Enabled = true;
 				button2.Enabled = true;
 				textBox1.Enabled = true;
@@ -96,26 +97,15 @@ namespace Admin_Module
 				MessageBox.Show(ex.Message, "Ошибка загрузки", MessageBoxButtons.OK, MessageBoxIcon.Error);
 			}
 		}
-		private void toolStripButton2_Click(object sender, EventArgs e)
-		{
-			dataGridView1.Visible = true;
-			dataGridView1.BringToFront();
-			toolStripButton3.Enabled = true;
-			toolStripButton2.Enabled = false;
-		}
-		private void toolStripButton3_Click(object sender, EventArgs e)
-		{
-			dataGridView1.Visible = false;
-			toolStripButton3.Enabled = false;
-			toolStripButton2.Enabled = true;
-		}
+
 
 		private void Form1_Load(object sender, EventArgs e)
 		{
 			label2.Enabled = false;
+			panel1.Visible = false;
 			dataGridView1.Visible = false;
-			toolStripButton2.Enabled = false;
-			toolStripButton3.Enabled = false;
+			btn_editor.Enabled = false;
+			btn_newquestion.Enabled = false;
 			button1.Enabled = false;
 			button2.Enabled = false;
 			textBox1.Enabled = false;
@@ -144,7 +134,7 @@ namespace Admin_Module
 			}
 			catch (Exception ex)
 			{
-				MessageBox.Show(ex.Message, "Ошибка", MessageBoxButtons.OK, MessageBoxIcon.Error);
+				MessageBox.Show(ex.Message, "Ошибка сохранения таблицы", MessageBoxButtons.OK, MessageBoxIcon.Error);
 			}
 
 		}
@@ -164,18 +154,17 @@ namespace Admin_Module
 			}
 			if (xlApp == null)
 			{
-				MessageBox.Show("Не удается найти excel!! Установите его или редактируйте таблицу в другом редакторе.");
 				Marshal.ReleaseComObject(xlApp);
 				GC.Collect();
 				GC.WaitForPendingFinalizers();
 				Cursor.Current = Cursors.Default;
-				return;
+				throw new Exception("Не удается найти Excel! Установите его или редактируйте таблицу в другом редакторе.");
 			}
 			Microsoft.Office.Interop.Excel.Workbook xlWorkBook;
 			Microsoft.Office.Interop.Excel.Worksheet xlWorkSheet;
 			try
 			{
-				
+
 				object misValue = System.Reflection.Missing.Value;
 				xlWorkBook = xlApp.Workbooks.Add(misValue);
 				xlWorkSheet = (Microsoft.Office.Interop.Excel.Worksheet)xlWorkBook.Worksheets.get_Item(1);
@@ -210,36 +199,48 @@ namespace Admin_Module
 				GC.WaitForPendingFinalizers();
 				Cursor.Current = Cursors.Default;
 			}
-			MessageBox.Show("Файл таблицы создан по адресу \"" + filename + "\"","Сохранение таблицы вопросов",MessageBoxButtons.OK,MessageBoxIcon.Asterisk);
+			MessageBox.Show("Файл таблицы создан по адресу \"" + filename + "\"", "Сохранение таблицы вопросов", MessageBoxButtons.OK, MessageBoxIcon.Asterisk);
 		}
 		private void button1_Click(object sender, EventArgs e)
 		{
-			dataGridView1.Sort(this.dataGridView1.Columns[0], ListSortDirection.Ascending);
-			StreamWriter fout = new StreamWriter(path, false);
-			string s = "";
-			int a = radioButton3.Checked ? 1 : 0;
-			s += a + ".1.";
-			a = radioButton2.Checked ? 1 : 0;
-			s += a + ".";
-			string lastChanged = DateTime.Now.ToString();
-			lastChanged = lastChanged.Replace(".", "").Replace(" ", "").Replace(":", "");
-			Text = s;
-			fout.Close();
+			try
+			{
+				DialogResult res = saveFileDialog1.ShowDialog();
+				if (res == DialogResult.OK)
+				{
+					string filename = saveFileDialog1.FileName;
+					Text = filename;
+					dataGridView1.Sort(this.dataGridView1.Columns[0], ListSortDirection.Ascending);
+					StreamWriter fout = new StreamWriter(path, false);
+					string s = "";
+					int a = radioButton3.Checked ? 1 : 0;
+					s += a + ".1.";
+					a = radioButton2.Checked ? 1 : 0;
+					s += a + ".";
+					string lastChanged = DateTime.Now.ToString();
+					lastChanged = lastChanged.Replace(".", "").Replace(" ", "").Replace(":", "");
+					Text = s;
+					fout.Close();
+				}
+				else throw new Exception("Файл не был сохранен!");
+			}
+			catch (Exception ex)
+			{
+				MessageBox.Show(ex.Message, "Ошибка сохранения файла", MessageBoxButtons.OK, MessageBoxIcon.Error);
+			}
+
+
+
+
 		}
 
 		private void toolStripButton4_Click(object sender, EventArgs e)
 		{
 
 		}
-
-		private void contextMenuStrip1_Opening(object sender, CancelEventArgs e)
-		{
-
-		}
-
 		private void dataGridView1_RowHeaderMouseClick(object sender, DataGridViewCellMouseEventArgs e)
 		{
-			if (e.Button==MouseButtons.Right)
+			if (e.Button == MouseButtons.Right)
 			{
 				if (MessageBox.Show("Вы действительно хотите удалить этот вопрос?", "Удаление вопроса", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) == DialogResult.Yes)
 				{
@@ -248,6 +249,177 @@ namespace Admin_Module
 					MessageBox.Show("Вопрос удален", "Удаление вопроса", MessageBoxButtons.OK, MessageBoxIcon.Information);
 				}
 			}
+		}
+
+		private void toolStrip1_ItemClicked(object sender, ToolStripItemClickedEventArgs e)
+		{
+
+		}
+
+		private void btn_editor_Click(object sender, EventArgs e)
+		{
+			dataGridView1.Visible = true;
+			panel1.Visible = false;
+			dataGridView1.BringToFront();
+		}
+
+		private void btn_main_Click(object sender, EventArgs e)
+		{
+			panel1.Visible = false;
+			dataGridView1.Visible = false;
+		}
+
+		private void btn_newquestion_Click(object sender, EventArgs e)
+		{
+			panel1.Visible = true;
+			dataGridView1.Visible = false;
+			panel1.BringToFront();
+		}
+
+		private void radioButton6_CheckedChanged(object sender, EventArgs e)
+		{
+			if (radioButton6.Checked == true)
+			{
+				label6.Visible = true;
+				numericUpDown2.Visible = true;
+			}
+			else
+			{
+				label6.Visible = false;
+				numericUpDown2.Visible = false;
+			}
+		}
+
+		private void numericUpDown3_ValueChanged(object sender, EventArgs e)
+		{
+			string s = numericUpDown3.Value.ToString();
+			textBox5.BackColor = Color.White;
+			textBox6.BackColor = Color.White;
+			textBox7.BackColor = Color.White;
+			textBox8.BackColor = Color.White; 
+			textBox9.BackColor = Color.White;
+			textBox10.BackColor = Color.White;
+			textBox5.ForeColor = Color.Black;
+			textBox6.ForeColor = Color.Black;
+			textBox7.ForeColor = Color.Black;
+			textBox8.ForeColor = Color.Black;
+			textBox9.ForeColor = Color.Black;
+			textBox10.ForeColor = Color.Black;
+			numericUpDown3.BackColor = Color.White;
+			numericUpDown3.ForeColor = Color.Black;
+			for (int i = 0; i < s.Length; i++)
+			{
+				switch (s[i])
+				{
+					case '1':
+						if (textBox5.Text != "")
+						{
+							textBox5.BackColor = Color.Green;
+							textBox5.ForeColor = Color.White;
+						}
+						else
+						{
+							numericUpDown3.BackColor = Color.Red;
+							numericUpDown3.ForeColor = Color.White;
+						}
+						break;
+					case '2':
+						if (textBox6.Text!="" && textBox5.Text != "")
+						{
+							textBox6.BackColor = Color.Green;
+							textBox6.ForeColor = Color.White;
+						}
+						else
+						{
+							numericUpDown3.BackColor = Color.Red;
+							numericUpDown3.ForeColor = Color.White;
+						}
+						break;
+					case '3':
+						if (textBox7.Text != "" && textBox6.Text != "" && textBox5.Text != "")
+						{
+							textBox7.BackColor = Color.Green;
+							textBox7.ForeColor = Color.White;
+						}
+						else
+						{
+							numericUpDown3.BackColor = Color.Red;
+							numericUpDown3.ForeColor = Color.White;
+						}
+						break;
+					case '4':
+						if (textBox8.Text != "" && textBox7.Text != "" && textBox6.Text != "" && textBox5.Text != "")
+						{
+							textBox8.BackColor = Color.Green;
+							textBox8.ForeColor = Color.White;
+						}
+						else
+						{
+							numericUpDown3.BackColor = Color.Red;
+							numericUpDown3.ForeColor = Color.White;
+						}
+						break;
+					case '5':
+						if (textBox9.Text != "" && textBox8.Text != "" && textBox7.Text != "" && textBox6.Text != "" && textBox5.Text != "")
+						{
+							textBox9.BackColor = Color.Green;
+							textBox9.ForeColor = Color.White;
+						}
+						else
+						{
+							numericUpDown3.BackColor = Color.Red;
+							numericUpDown3.ForeColor = Color.White;
+						}
+						break;
+					case '6':
+						if (textBox10.Text != ""&&textBox9.Text != "" && textBox8.Text != "" && textBox7.Text != "" && textBox6.Text != "" && textBox5.Text != "")
+						{
+							textBox10.BackColor = Color.Green;
+							textBox10.ForeColor = Color.White;
+						}
+						else
+						{
+							numericUpDown3.BackColor = Color.Red;
+							numericUpDown3.ForeColor = Color.White;
+						}
+						break;
+					default:
+						numericUpDown3.BackColor = Color.Red;
+						numericUpDown3.ForeColor = Color.White;
+						break;
+				}
+			}
+		}
+
+		private void textBox5_TextChanged(object sender, EventArgs e)
+		{
+			numericUpDown3_ValueChanged(sender, e);
+
+		}
+
+		private void textBox6_TextChanged(object sender, EventArgs e)
+		{
+			numericUpDown3_ValueChanged(sender, e);
+		}
+
+		private void textBox7_TextChanged(object sender, EventArgs e)
+		{
+			numericUpDown3_ValueChanged(sender, e);
+		}
+
+		private void textBox8_TextChanged(object sender, EventArgs e)
+		{
+			numericUpDown3_ValueChanged(sender, e);
+		}
+
+		private void textBox9_TextChanged(object sender, EventArgs e)
+		{
+			numericUpDown3_ValueChanged(sender, e);
+		}
+
+		private void textBox10_TextChanged(object sender, EventArgs e)
+		{
+			numericUpDown3_ValueChanged(sender, e);
 		}
 	}
 }
